refactor events, conditions and transitions package
A new package 'besser.agent.core.transition' contains all transition-related classes. Specific instances of events and conditions have been moved to besser.agent.library.transition package.

To be decided if we move github and gitlab events to the library package as well.
