--- besser/agent/library/event/event_template.py
@@ -1,12 +1,0 @@
-from besser.agent.core.session import Session
-
-
-def event_template(session: Session, event_params: dict) -> bool:
-    """
-    Used as a reference for event method signatures.
-
-    Args:
-        session (Session): the current user session
-        event_params (dict): the event parameters
-    """
-    pass
